UpdateTest and Delete Project

+++ src/Services/Workspace/src/Application/Workspace/Commands/DeleteWorkspace/DeleteWorkspaceHandler.cs
@@ -1,4 +1,5 @@
-﻿using Application.Common.Interfaces;
+﻿using Application.Common.Exceptions;
+using Application.Common.Interfaces;
 using Application.Repositories;
 using MediatR;
 using System;
@@ -21,7 +22,15 @@ namespace Application.Workspace.Commands.DeleteWorkspace
         }
         public async Task<DeleteWorkspaceReturnDto> Handle(DeleteWorkspaceCommad request, CancellationToken cancellationToken)
         {
-            return await _workspaceRepository.DeleteAsync(request.WorkspaceId, _currentUserService, cancellationToken);
+            //var exist = await _workspaceRepository.GetAsync(request.WorkspaceId);
+            if(request.WorkspaceId == Guid.Empty)
+            {
+                throw new BusinessRuleException($"Workspace {request.WorkspaceId} doesn't exist");
+            }
+
+            var dto = await _workspaceRepository.DeleteAsync(request.WorkspaceId, _currentUserService, cancellationToken);
+
+            return dto;
         }
     }
 }
